Begin extracting TestDB testing infrastructure
