+++ testdb_test.go
@@ -1,0 +1,10 @@
+package pgxschema
+
+import "github.com/ory/dockertest"
+
+type ConnInfo struct {
+	DockerRepo string
+	DockerTag  string
+	DSN        string
+	Resource   *dockertest.Resource
+}
